Implement user-consented Zipalign error handling
This commit introduces error handling for `zipalign` failures during APK payload generation. In case of a `zipalign` alignment error, the user is now prompted to run `ZipalignReinstallTool.py` for resolving the issue. The script execution is contingent upon user consent, enhancing the process's transparency and control. This fix addresses potential disruptions in payload generation due to `zipalign` misconfigurations or failures.